Create Redis migration app
This changeset introduces a bare-bones script to run as a quick app for migrating Redis keys from a source instance to a destination instance, and then watching for new write commands (SET or SETEX) in the source instance so that they can be copied to the destination instance.

This will allow us to update an app that needs to point to a new Redis instance without any disruption in service.